Added General Test all cases- Error, Transparency, boundingBoxes, IgnoredBoxes
