+++ package.json
@@ -9,10 +9,7 @@
   "main": "index.js",
   "scripts": {
     "test": "node test.js && node multiple-url-sample.js",
-    "ignoredBox": "cd tests/generalCases && node ignoredBoxes.js",
-    "boundingBox": "cd tests/generalCases && node boundingBoxes.js",
-    "error": "cd tests/generalCases && node errorSettings.js",
-    "transparency": "cd tests/generalCases && node transparency.js"
+    "generalTest": "cd tests/generalCases && node index.js"
   },
   "repository": {
     "type": "git",
